requirements.txt
# Excel processing libraries
openpyxl>=3.1.0
pandas>=2.0.0
xlrd>=2.0.1

# Date and time handling
jdatetime>=4.1.0

# Data manipulation
numpy>=1.24.0

# Optional: For better Excel support
xlsxwriter>=3.1.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# django
Django==4.2.23
------

django-extensions
djangorestframework
django-htmx
